+++ Runtime/Chat/Models/ChatMessage.cs
@@ -1,6 +1,6 @@
 using Firebase.Firestore;
 
-namespace SocialManager.Chat.Models
+namespace Suhdo.FSM.Chat.Models
 {
     [FirestoreData]
     public class ChatMessage